add an authentication proxy
This commit adds authentication by starting tensorboard in a new
network namespace (by way of unshare).  In that same namespace it
starts a proxy based off of work done by stanford (license and info
included in this commit). slirp4netns then creates a bridge connecting
the outside world to that proxy - thus allowing for authentication and
isolation.